backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.6.1
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
chromadb==0.4.22
pandas
httpx==0.25.2
pydantic==2.5.3
python-multipart==0.0.6
tiktoken==0.5.2
beautifulsoup4==4.12.2
playwright==1.40.0
